Added xlink namespace
I had a lot of troubles with some SVG files which has `<image>`tags with xlink base64 data inside. Without adding the xlink namespace, the SVG sprite is not valid and PhantomJS can't get the SVG dimension. If you try to open the SVG sprite on Chrome, you will get a message like this one:

*error on line 1 at column 332: Namespace prefix xlink for href on image is not defined*