Output padded items offset the correct amount
Apologies, I kind of screwed up in my last pull request. What I previously did was add extra height and width to the items size to account for the padding. This obviously made items bigger than they should be.

Now I've removed that and just offset the background position by the padding amount. This means that the item's size stays the same as the original image.