Output sprite css file to multiple locations
This allows the user to output the sprite file to multiple locations.

For instance, I have it set as `cssFile: ["css/sprites.scss", "scss/components/_sprites.scss"]`, allowing for a preview compatible CSS file, and an scss file which integrates with my Sass system. The first array item is submitted to the preview document.

Admittedly, it's a niche feature, but I find it very handy.