Include viewbox in code preview
Currently, if viewbox is not included, the parent SVG is at default height, even if width is specified in CSS (downscaling an icon proportionnaly by specifying only the width becomes impossible). I think the spec of SVG2 will remove this requirement, but right now latest Chrome (62.0.3202.94) and Firefox Quantum (58) don't do it, so it's impossible to manage an SVG icon this way.